refactor: SRP 원칙에 따른 클래스 분리에 맞춰 테스트 클래스도 분리
