+++ src/test/java/com/nexus/sion/feature/squad/command/SquadCommandIntegrationTest.java
@@ -96,7 +96,7 @@ class SquadCommandIntegrationTest {
             .startDate(LocalDate.of(2025, 1, 1))
             .expectedEndDate(LocalDate.of(2025, 12, 31))
             .numberOfMembers(1)
-            .budget(10_000_000L)
+            .budget(new BigDecimal(10_000_000L))
             .status(ProjectStatus.WAITING)
             .requestSpecificationUrl("http://example.com/spec")
             .domainName("CS")
