+++ src/test/java/com/nexus/sion/feature/squad/command/application/service/SquadLifeCycleImplTest.java
@@ -1,0 +1,7 @@
+package com.nexus.sion.feature.squad.command.application.service;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+class SquadLifeCycleImplTest {
+
+}
